requirements.txt
Django==3.2.9
python-dotenv==0.19.2
asgiref==3.8.1
gunicorn==23.0.0
packaging==24.1
sqlparse==0.5.1
whitenoise==6.7.0
uvicorn==0.15.0
python-decouple==3.8
openai==0.27.0
psycopg2-binary==2.9.3